requirements.txt
requests
m3u8
pycryptodome